# api_documentation.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Basic Bank System - OpenAPI 3.0",
    "description": "Dokumentasi API Basic Bank System\n\n- [Repository of Basic Bank System](https://github.com/yusuftri023/Basic-Banking-System)",
    "contact": {
      "email": "yusuftri023@gmail.com"
    },
    "version": "1.0.1"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
...
  "paths": {
...
    "/api/v3/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Get Users lists",
        "description": "Get Users Lists",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Pagination display limit",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Request has succeeded"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_name": {
                          "type": "string",
                          "example": "Hendra Gunawan"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "Failed operation should have been 200 but open api won't let response code duplicate exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Data Empty"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Add a new user to database",
        "description": "Add a new user to database",
        "operationId": "Users",
        "requestBody": {
          "description": "Create a new Users in the website",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_name": {
                    "type": "string",
                    "example": "Hendra Gunawan"
                  },
                  "phone_number": {
                    "type": "string",
                    "example": 8887329342
                  },
                  "identity_type": {
                    "type": "string",
                    "example": "KTP"
                  },
                  "identity_number": {
                    "type": "string",
                    "description": "Serial number of identity card",
                    "example": 779288223794
                  },
                  "address": {
                    "type": "string",
                    "description": "address",
                    "example": "Bali"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 8887329342
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully added to the database"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_name": {
                          "type": "string",
                          "example": "Hendra Gunawan"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "One of inputted data already exist in database and need to be unique",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get Users details by userid",
        "description": "Get Users details by userid",
        "operationId": "getUsersById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully added to the database"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_name": {
                          "type": "string",
                          "example": "Hendra Gunawan"
                        },
                        "profiles": {
                          "type": "object",
                          "properties": {
                            "profile_id": {
                              "type": "integer",
                              "example": 1
                            },
                            "user_id": {
                              "type": "integer",
                              "example": 1
                            },
                            "phone_number": {
                              "type": "string",
                              "example": 8887329342
                            },
                            "identity_type": {
                              "type": "string",
                              "example": "KTP"
                            },
                            "identity_number": {
                              "type": "string",
                              "description": "Serial number of identity card",
                              "example": 779288223794
                            },
                            "address": {
                              "type": "string",
                              "description": "address",
                              "example": "Bali"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update the existing user data",
        "description": "Update the existing user data",
        "operationId": "putUsersById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          }
        ],
        "requestBody": {
          "description": "Create a new Users in the website",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_name": {
                    "type": "string",
                    "example": "Hendra Gunawan"
                  },
                  "phone_number": {
                    "type": "string",
                    "example": 8887329342
                  },
                  "identity_type": {
                    "type": "string",
                    "example": "KTP"
                  },
                  "identity_number": {
                    "type": "string",
                    "description": "Serial number of identity card",
                    "example": 779288223794
                  },
                  "address": {
                    "type": "string",
                    "description": "address",
                    "example": "Bali"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "succesfully update user data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "User data succesfully updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "string",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Data not found in database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "string",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Data not found in database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "string",
                      "example": 409
                    },
                    "message": {
                      "type": "string",
                      "example": "data already exist"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete a user from database",
        "description": "Delete a user from database",
        "operationId": "deleteUsersById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "user id",
            "required": true,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "code": {
                      "type": "string",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Users successfully deleted"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "user_name": {
                          "type": "string",
                          "example": "Hendra Gunawan"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "code": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User does not exist"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie"
      }
    }
  }
}